Rename idle time unit to idle timeout unit
For consistency with:

- idle timeout property in the same data object
- connection level properties
