+++ vertx-pg-client/src/test/java/io/vertx/pgclient/PgPoolTest.java
@@ -19,14 +19,23 @@ package io.vertx.pgclient;
 
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
+import io.vertx.ext.unit.junit.Repeat;
+import io.vertx.ext.unit.junit.RepeatRule;
 import io.vertx.sqlclient.PoolOptions;
-import java.util.concurrent.atomic.AtomicReference;
+import org.junit.Rule;
 import org.junit.Test;
+
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
  */
 public class PgPoolTest extends PgPoolTestBase {
+
+  @Rule
+  public RepeatRule rule = new RepeatRule();
 
   @Override
   protected PgPool createPool(PgConnectOptions options, PoolOptions poolOpts) {
@@ -161,5 +170,29 @@ public class PgPoolTest extends PgPoolTestBase {
       pool.close();
     }
   }
-}
 
+  @Repeat(500)
+  @Test
+  public void checkBorderConditionBetweenIdleAndGetConnection(TestContext ctx) {
+    int concurrentRequestAmount = 100;
+    int idle = 1000;
+    int poolSize = 5;
+
+    options.setIdleTimeout(idle).setIdleTimeoutUnit(TimeUnit.MILLISECONDS);
+    poolOptions.setMaxSize(poolSize).setIdleTimeout(idle).setIdleTimeoutUnit(TimeUnit.MILLISECONDS);
+    PgPool pool = createPool(options, poolOptions);
+
+    Async async = ctx.async(concurrentRequestAmount);
+    for (int i = 0; i < concurrentRequestAmount; i++) {
+      CompletableFuture.runAsync(() -> {
+        pool.query("SELECT CURRENT_TIMESTAMP;").execute(ctx.asyncAssertSuccess(rowSet -> {
+          pool.query("select count(*) as cnt from pg_stat_activity where application_name like '%vertx%' and state = 'active'").execute(ctx.asyncAssertSuccess(rows -> {
+            Integer count = rows.iterator().next().getInteger("cnt");
+            ctx.assertInRange(count, 1, poolSize, "Oops!...Connections exceed poolSize. Are you leaked connections?.");
+            async.countDown();
+          }));
+        }));
+      });
+    }
+  }
+}
