+++ vertx-sql-client/src/main/java/io/vertx/sqlclient/impl/PoolBase.java
@@ -24,6 +24,7 @@ import io.vertx.sqlclient.SqlConnection;
 import io.vertx.sqlclient.Transaction;
 import io.vertx.sqlclient.impl.command.CommandBase;
 import io.vertx.sqlclient.impl.command.CommandResponse;
+
 import java.util.concurrent.TimeUnit;
 
 /**
@@ -47,7 +48,7 @@ public abstract class PoolBase<P extends PoolBase<P>> extends SqlClientBase<P> i
       throw new IllegalArgumentException("Pool max size must be > 0");
     }
     this.context = context;
-    long idleTimeOut = TimeUnit.MILLISECONDS.convert(options.getIdleTimeout(), options.getIdleTimeUnit());
+    long idleTimeOut = TimeUnit.MILLISECONDS.convert(options.getIdleTimeout(), options.getIdleTimeoutUnit());
     this.pool = new ConnectionPool(context, this::connect, maxSize, options.getMaxWaitQueueSize(), idleTimeOut);
     this.closeVertx = closeVertx;
   }
